Increase DownloadModuleSlice read buffer.
This is a temporary change, and we should look into upstreaming a
proper fix.

Since all reads are done in series, the network overhead for sending
thousands of small requests adds up. By increasing the read buffer size
there is a 10x improvement in transfer speed, and 30x improvement in
transfer speed over port forwarding for a 16MB file.