Support for template instantiation in expression evaluator
To get the templates from DWARF, we simply assume that a symbol
containing '<' is a template.

We also introduce a callback from Clang to lookup template
instantiation lazily. Unfortunately, we still do parse all
the template's instantiations from DWARF (this only takes
significant amount of time - several seconds for couple of
thousand instantiations) on the first evaluation.