jThreadsInfo returns stack memory
This patch adds parts of the stack that should be useful for unwinding
to the jThreadsInfo reply from lldb-server. We return the top of the
stack (12 words), and we also try to walk the frame pointer linked list
and return the memory containing frame pointer and return address pairs.
The idea is to cover the cases with and without frame pointer omission.

Differential Revision: https://reviews.llvm.org/D74398

This patch landed upstream and later reverted because of a bug
in process' L1 memory cache. The fix to the bug was proposed in
https://reviews.llvm.org/D77765, but not accepted because of
disagreements on testing.