Lookup unwind row from unwind info base
If the unwind info base address does not agree with the symbol start
address, lldb can currently find a wrong unwind table row because it
computes the offset from the symbol start address.

With this patch we adjust the offset for unwind row lookup to start from
unwind info base (if both the unwind info base and function start are
available).

We also test this on a simple example where the unwind info starts at an
address after function start. Similar situation also occurs with mingw's
handling of the ms_hook_prologue attribute. There, the unwind info
starts before the function low-pc address. As a result, we see lldb not
being to able unwind stacks with hotpatchable functions (i.e., functions
with ms_hook_prologue).