Derive core file modules from nt_file note
This introduces two changes:
1. Infer module list from NT_FILE section.
2. Read build-ids from the first page of each module.

The advantage of (1) is that core files do not need
the main executable to be able to list the modules.
Another advantage is an easier implementation of (2).

There is also a disadvantage - if some modules do not
have their first page in the core file, those would
be missing from the module list. This only applies
to the modules there have not been mapped to physical
memory, so this should not be a big deal. In my tests
with Unreal crash dumps from Stadia, all modules have
their first pages in the core file. I have only seen
missing pages locally, those pages came from libraries
that were not used by the executable.

This new behaviour can be disabled by changing the LLDB
setting that this CL introduces:

```
settings set plugin.process.elf-core.load-modules-from-core-note false
```

If loading modules from nt_file note is disabled or if
the core file does not contain an nt_file note, LLDB
will fall back to the original method of loading modules
via the POSIX dynamic linker plugin.