Minidump should contain just enough info
Full memory is not needed in the minidump and populating full
data takes significant amount of time which should be avoided.
This change limits a number of bytes saved per thread.
Additionally we fix module names (for remote runs it was taking
cached values from the local lldb). This is not upstreamed yet.